+++ lib/src/actions/base_actions.dart
@@ -5,3 +5,4 @@ part 'actions/explosion.dart';
 part 'actions/hands.dart';
 part 'actions/immortality.dart';
 part 'actions/points.dart';
+part 'actions/transfer.dart';