+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.2.10
+- Variant: Andernach Chess (capturing pieces change colour).
+- `ActionTransferOwnership`: changes the colour of the moving piece.
+
 ### 1.2.9
 - Fixed a bug with lame leaper squares being calculated incorrectly.
 - Fixed a bug where points-based games could result in draws when the opponent would win regardless next round. `GameResult.soft` defines result types that won't delegalise moves.